workable implementation lasers
Everything functions now.
Three issues remain:
The static light sometimes just doesn't activate a laser. No idea why. This should just be ignorable
The other two are both caused by the laser raycast now ignoring the wall when it's disappeared (which is necessary to detect if a laser should pass through it):
- the wall does not return when the laser color is changed.
- the wall does not return when the laser is blocked.
Both of these are unsolvable in the current implementation. The blocking being impossible in particular is actually a pretty significant issue.