Moving platforms implemented
Implemented the moving platforms feature. This also necessitated changing the method by which death is checked. The difference is that we will now need to also place an appropriate area2D over the spikes tilemap to make the feature work (it was not possible to walk through them in the previous implementation).
Also a couple of (inactive) methods removed from the lasers that I guess github hadn't caught up to yet last time.