full mirror implementation; general fixes
Fully implemented working mirrors, that can be rotated dynamically (press G ingame to test).
Generally stress-tested the lasers for this, and nothing seems to break significantly. When multiple lasers affect the same block, they mix based on whichever happened to hit the most recently; this seems fine to me. It is not currently possible to collect information for multiple lasers hitting a block at the same time. This also means that when you block the path for just one laser, the target block will always disable itself entirely. This seems fine to me, we should avoid this situation anyway.
Also changed the moving platform collision to be a little smaller. Mind that its Area2D needs to extend *beyond* that of the deathzone, even if that's by just one pixel.
Added guidelines for the lasers that are only visible in the editor window.
Tentatively implemented static lights affecting eachother. However, I am unable to properly test this functionality without being able to rotate/move lights into eachother, so decided to wait before ensuring it actually works this way.