button implementation; general fixes
Buttons are implemented. Change their position and rotation per instance in the inspector window. Rotation Value is the amount by which the button rotates the light.
The lights operate with a new control node because the position of the button needs to remain fixed while the light rotates. The button is a child because it was the simplest way to associate instances.
New sprites for buttons and static lights are placeholder, just here to make sure we definitely have something for Monday. The same applies to the static light that activates within range of the button.
Changed the objects from alternating visible property to enabled. Doesn't really matter.
Changed the check for whether a shapecast is no longer colliding with an object to move outside its own collision check. This should be what was causing the delay for static lights returning to their normal states, but the problem actually still remains. It does certainly make the code more stable. We should only seriously look at the delay after sorting out the tween anyway.