changed spike texture to match tilesize
this completely destroyed my level due to some strange scale wizardry. everything here is half the size it should be but I can't fully figure out why. Should likely not pose a problem when we just make a new scene, though. It might have something to do with there having been two tilemapcontrols (which was an accident)